Add files via upload
Continuing from our group coursework, we were assigned to each implement a version of the
automaton. This was the NFA before and after the conversion to a DFA. I implemented the NFA
using sequential searching. The other version was a parallel search, followed by the DFA
implementations using direct encoding and data structure initialization.